+++ hadoop-ozone/ozone-manager/src/test/java/org/apache/hadoop/ozone/om/request/om_target/TestOMTarget.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.hadoop.ozone.om.request.target;
+package org.apache.hadoop.ozone.om.request.om_target;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
