+++ hadoop-ozone/ozone-manager/src/test/java/org/apache/hadoop/ozone/om/request/om_target/TestOMTargetAddRequest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.hadoop.ozone.om.request.target;
+package org.apache.hadoop.ozone.om.request.om_target;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
